refactor(evals): extract assertion recorder
